[Bronze V] Title: 이상한 기호, Time: 48 ms, Memory: 5496 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -3,7 +3,7 @@ using System.Xml;
 int[] inputArray = Array.ConvertAll(Console.ReadLine().Split(' '), int.Parse);
 
 
-long Calculate(int A, int B)
+long Calculate(long A, long B)
 {
     long output = (A + B) * (A - B);
     return output;
